requirements.txt
streamlit==1.0.0
pandas==2.0.3
openpyxl==3.0.10
PyGithub==2.4.0
python-dotenv==1.0.1